Use the new `create` endpoint from the bits service
Instead of using a `PUT` request, CC will upload the blob as a `POST`
and without any guid. The bits-service will store it, create an guid and
return it to the CC. This created guid should then be used to perform
any operations against the blob and the blobstore in bits-service.

[#112890545]

Signed-off-by: Tiago Scolari <tscolari@pivotal.io>